# openapi8-8-22.yaml
openapi: 3.0.2
servers:
  - url: http://localhost:8080
info:
  description: Esta es la API que vamos a crear en clase
  version: "1.0"
  title: Delivery API
...
paths:
...
  /restaurants:
     get:
      tags: 
        - restaurants
      summary: Listado de restaurants
      description: Devuelve el listado completo de todas los restaurantes registrados. Es posible filtrar por la categoria del pedido
      operationId: getRestaruants
      parameters:
        - name: category
          in: query
          description: Dsitancia de que se encuentra entre el punto de entrga y el restaurante
          required: false
          schema:
            type: string
            default: "Mexicano"
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Restaurant'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        '404':
          description: No se ha encontrado restaurante con esa categoria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
              example:
                internalError: 1
                message: "No se ha podido encontrar el restaurante"
     post:
      tags:
        - restaurants
      summary: Registrar un restaurante
      description: Registra un nuevo restaurante
      operationId: postRestaurnat
      requestBody:
        description: Los datos de un nuevo resaturante
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestaurantIn'
      responses:
        '201':
          description: Se ha registrado el restaurnate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        '400':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
  /restaruant/{restaurantId}:
    get:
      tags: 
        - restaurants
      summary: Obtiene los datos de un restaurante
      description: Devuelve los datos de un restaurante determinado
      operationId: getRestaurnat
      parameters:
        - name: restaurantId
          in: path
          description: Identificador del restaurante
          required: true
          schema:
            type: number
            format: int64
            example: 1
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        '404':
          description: No se ha encontrado el restuarante
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
              example:
                internalError: 1
                message: "No se ha podido encontrar el restaurante"
    delete:
      tags: 
        - restaurants
      summary: Da de baja un restaurante
      description: Da de baja un restaurante determinado
      operationId: deleteRestaurant
      parameters:
        - name: restaurantId
          in: path
          description: Identificador del restaurnate
          required: true
          schema:
            type: number
            format: int64
            example: 23
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        '404':
          description: No se ha encontrado el restaurante
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
              example:
                internalError: 1
                message: "No se ha podido encontrar el restaurante"
    put:
      tags: 
        - restaurants
      summary: Modifica un restaurante
      description: Modifica los datos de un restaurante determinado
      operationId: modifyRestaurant
      parameters:
        - name: restaurantId
          in: path
          description: Identificador del restaurante que se va a modificar
          required: true
          schema:
            type: number
            format: int64
            example: 23
      requestBody:
        description: Los datos nuevos del restaurante
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestaurantIn'
      responses:
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestaurantIn'
        '404':
          description: No se ha encontrado el restaurante
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
              example:
                internalError: 1
                message: "No se ha podido encontrar el restaurante"
                
                
    patch:
      tags: 
        - restaurants
      summary: Modifica el estado de un restaurante
      description: Modifica el estado del restaurante si esta abierto o cerrado
      operationId: patchRestaurnat
      parameters:
        - name: restaurantId
          in: path
          description: boolean que dependera del estado del local
          required: true
          schema:
            type: boolean
            format: int64
            example: 23
      requestBody:
        description: Los datos nuevos del estado del local
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestaurantIn'
      responses:
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        '404':
          description: No se ha encontrado el restaurante
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
              example:
                internalError: 1
                message: "No se ha podido encontrar el restaurante"
...
  /rider/{riderId}:
    get:
      tags: 
        - riders
      summary: Obtiene los datos de un rider
      description: Devuelve los datos de un rider determinado
      operationId: getRider
      parameters:
        - name: riderId
          in: path
          description: Identificador del rider
          required: true
          schema:
            type: number
            format: int64
            example: 1
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rider'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        '404':
          description: No se ha encontrado el rider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
              example:
                internalError: 1
                message: "No se ha podido encontrar el rider"
    delete:
      tags: 
        - riders
      summary: Da de baja un restaurante
      description: Da de baja un restaurante determinado
      operationId: deleteRider
      parameters:
        - name: riderId
          in: path
          description: Identificador del rider
          required: true
          schema:
            type: number
            format: int64
            example: 23
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rider'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        '404':
          description: No se ha encontrado el rider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
              example:
                internalError: 1
                message: "No se ha podido encontrar el rider"
    put:
      tags: 
        - riders
      summary: Modifica un rider
      description: Modifica los datos de un rider determinado
      operationId: modifyRider
      parameters:
        - name: riderId
          in: path
          description: Identificador del rider que se va a modificar
          required: true
          schema:
            type: number
            format: int64
            example: 23
      requestBody:
        description: Los datos nuevos del rider
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiderIn'
      responses:
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rider'
        '404':
          description: No se ha encontrado el rider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
              example:
                internalError: 1
                message: "No se ha podido encontrar el rider"
                
                
    patch:
      tags: 
        - riders
      summary: Modifica el nombre del rider
      description: Modifica el nombre del rider
      operationId: patchRider
      parameters:
        - name: riderId
          in: path
          description: string que cambia el nombre del rider
          required: true
          schema:
            type: string
            example: "Daniel"
      requestBody:
        description: Los datos nuevos del rider
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiderIn'
      responses:
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rider'
        '404':
          description: No se ha encontrado el rider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
              example:
                internalError: 1
                message: "No se ha podido encontrar el rider"
components:
  schemas:
...
    RestaurantIn:
      required:
      - name
      - address
      - capacity
      - operative
      - mediumPrice
      - category
      properties:
        name:
          type: string
          example: "La zarzuela"
        address:
          type: string
          example: "plaza Aragon n8"
        capacity:
          type: integer
          format: int32
          example: 50
        operative:
          type: boolean
          example: true
        mediumPrice:
          type: number
          format: double
          example : 30.5
        category:
          type: string
          example : "Asiatico"
          
      type: object
    Restaurant:
      allOf:
        - type: object
          properties:
            id:
              type: integer
              format: int64
              example: 23
        - $ref: '#/components/schemas/RestaurantIn'
    RiderIn:
      properties:
        name:
          type: string
          example: "Jose Luis"
        surname:
          type: string
          example: "Lopez"
        vehicle:
          type: string
          example: "bicicleta"
        maxSpeed:
          type: number
          format: int32
          example: 50
      type: object
    Rider:
      allOf:
        - type: object
          properties:
            id:
              type: integer
              format: int64
              example: 23
        - $ref: '#/components/schemas/RiderIn'       
...
    ErrorResponse:
      properties:
        internalError:
          type: number
          format: int32
        message:
          type: string
      type: object
    ResourceNotFound:
      allOf:
        - type: object
        - $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      allOf:
        - type: object
        - $ref: '#/components/schemas/ErrorResponse'
      example:
        internalError: 999
        message: "Internal Server Error"
    BadRequest:
      allOf:
        - type: object
        - $ref: '#/components/schemas/BadRequest'
      example:
        internalError: 400
        message: "BadRequest"